resolve: process time zone switching logic in middleware
