+++ src/bk-user/bkuser/apis/web/personal_center/views.py
@@ -11,7 +11,6 @@ specific language governing permissions and limitations under the License.
 
 from typing import Dict
 
-from django.conf import settings
 from django.utils.translation import gettext_lazy as _
 from drf_yasg.utils import swagger_auto_schema
 from rest_framework import generics, status
@@ -222,8 +221,6 @@ class TenantUserTimeZoneUpdateApi(ExcludePatchAPIViewMixin, generics.UpdateAPIVi
         tenant_user = self.get_object()
         tenant_user.time_zone = data["time_zone"]
         tenant_user.save(update_fields=["time_zone", "updated_at"])
-
-        request.session[settings.TIME_ZONE_SESSION_KEY] = data["time_zone"]
 
         return Response(status=status.HTTP_204_NO_CONTENT)
